PY-23842 Don't pollute the cache of package managers with temporary ones
Otherwise, we end up having stale instances of PyPackageManager
referencing fake SDK with an empty name that is reused
every time we try to install setuptools and pip in a new virtualenv
(since the name of an SDK is used as the key in that cache now).
As a result, we kept installing packaging tools to the first virtualenv
created during the current IDE session.

(cherry picked from commit 38ef3b3)