[Kotlin] Temporary workaround for UastContext CCEs
Kotlin plugin 1.1.0 bundled UAST, whereas 1.1.2 and later don't.
-- UAST has been pushed into the platform (Intellij). When we
started bundling Kotlin in Studio, this scenario became possible:
 - user installed 1.1.0 or earlier in local config/, so there exists a
Kotlin/lib/uast-common.jar.
 - Studio 3.0 ships with its own bundled version, which wins over the
older one (left untouched, in the local config folder, possibly
even imported from an older config)
 - Studio allows unbundled plugin updates; a new Kotlin plugin is
unzipped on top of the locally installed one. Studio checks whether to
remove the current plugin, but at the time of the check, the current
plugin is the bundled one, and we never allow these plugins to be
deleted. So uast-common.jar survives the update.
 - we now have two versions of UAST, one loaded by the platform and
one by the plugin classloader.

In this patch, we manually exclude UAST when loading an unbundled
Kotlin plugin, which must be recent enough to win the version check.

Bug: 63525614
Test: N/A
Change-Id: I7f0f6ec7752da58ff1d0c026736d6d54ed4c66b5