Extends IdeaModuleBase to support jars
IdeaModuleBase already extends a file set that supports
zips or directories as inputs. This changes it so if the
output of the model is a .jar, it will use the zip code
path. It's unfortunate that the ant base class has two
variables and expects only one of them to be set.