[gui-test] speedup cell readers
We created a new Robot instance (and dispose resources after each usage before) for each getValue() operation to read a cell value. It costs about 1 second on Windows, that could be very expensive to calculate path for a tree. Here we are using simple DFS algorithm to iterate between children and other grand children of current container.