IDEA-175622 Revert "lst: do not mark file dirty"
This is necessary to refresh CLM if document was modified, but
changes were later reverted (without using Undo).
In this case, VcsDirtyScopeVfsListener will not be notified,
because no changes are saved on disk.
And FileStatusManagerImpl will not mark file dirty, because
document has different modification stamp
(and FileDocumentManagerImpl.isFileModified returns `true`).

This reverts commit 1906673cbee6c735d1be57d3bc5086a4b121f0cc.