PY-25491 Use AST-based search to find 'as' target expression
We don't store stubs for 'as' targets of type PyTargetExpression in
PyImportElements. We store the 'as' name in the stubs for
PyImportElements instead. The getAsName() checks if the element has
a stub, and if it hasn't, then asks its AST about the name. It used to
access AST via getStubOrPsiChild() and it was always AST because it
happened after a 'getStub() == null' check. But now we have green stubs
so getStubOrPsiChild() accesses the green stub via getGreenStub() even
though we have AST.

(cherry picked from commit 6291058)

(cherry picked from commit 9e54b18)