Add name ID for the usage tree component
The UI tests benefit from having the name set on the usage
tree component. The name acts as an identifier for UI tests.
By using an identifier, locating a particular UI component
is easy to program. If the name was not set, the UI tests
would attempt to identify the tree by checking the tree's
parent's and the tree's class types, which is less simple
and less clear than using an identifier.

Test: UnusedResourceEvaluationTest
Change-Id: I9bc8f3cc67d2f09679a2d936504b9f74c1340194