[UITests] Implement JUnit's "perTest" forked test execution
We can't use JUnit's forkMode attribute directly because the UI testing
framework is using BootstrapUITests as the entry point, which to JUnit
looks like an opaque test class. (BootstrapUITests was a fix for the
classpath-too-long problem, similar to BootstrapTests used for unit
tests; we should remove both of these eventually).

The custom Ant task is therefore in charge of splitting up a test
group requiring this execution style into individual test classes and
scheduling them as separate JUnitTasks.

Bug: 37144715
Test: manual
Change-Id: I4f9d3d43991daf45873df8deb2b555901d10de53