IDEA-168392 Don't warn about rebasing a merge which had no conflicts
Rebasing over a merge is a normal procedure, and the only annoying thing which may happen is resolving conflicts again, if they happened during the previous merge.

Therefore the warning should be shown only if the user tries to rebase over a merge with conflicts.

Additionally, the warning itself is made a bit less frightening.