[Bug: 65100260] Use AllIcons.Nodes.Module instead of ANDORID_MODULE
StudioIcons.Shell.Filetree.ANDROID_MODULE can not be used as an icon for
com.intellij.openapi.fileChooser.actions.GotoModuleDirectory because
ANDROID_MODULE belongs to artwork module that is inside Android plugin and
is not accessible from core classloader.

This CL partially reverts ag/2742348 to not use ANDROID_MODULE icon for
actions in RichPlatformActions.xml

Test: Manually confirmed that no error is caused while searching for Module
Directory action.

Change-Id: Ibaa8f1691fc9852c9133e76ba25740c8a0cd7b16