Don't mark py2only directory in python-helpers as source root
Sometimes its content is copied literally to the root of the build
output of "python-helpers" module, and then this Python 2 specific
version of the library is used unconditionally (as the first in sys.path)
in docstring_formatter.py script that might be running on Python 3
interpreter. It may lead to docstrings not being rendered properly in
PyCharm launched from IDEA project for no clear reason.