Better defaults for AndroidJUnit run configurations
For compatibility with the Gradle test runner, we modify the working
directory and default JVM flags. This only takes effect in newly created
projects unfortunately, but there's only so much we can do about it. In
PostSyncProjectStep we overwrite the working directory of template run
configurations but we cannot overwrite the JVM flags, because users may
actually have useful JVM flags in their templates.

Test: UnitTestingSupportIntegrationTest
Change-Id: I67ee5b84a20d5a2cebd23b3cfc9866606e565dcf