IDEA-81364 Read GitConfig options case insensitively
ini4j allows to convert all options to the lower case.

An issue still remains with the case sensitivity of section names
(e.g. [REMOTE "origin"] won't be parsed), and ini4j doesn't help with `setLowerCaseSection` here, since the latter would convert the remote name to the lower case as well (thus breaking `[remote "GITHUB"]`), but it seems to be an exotic use case which can be handled later.