[gui-test] introduce GuiTestRule; reload IDE after fatal errors
The reason that we have introduce GuiTestRule (and dropped GuiTestBase class) is in a low accuracy of screenshots on failures. Previously we could register a welcome frame step on a screenshot on failures while an exception has been raised inside IdeFrame. It could be solved only by switching from the traditional @Before setUp() @After tearDown() model to the rules model of JUnit 4. The GuiTestRule was inspired by AndroidStudio twin class and enlarged by a FatalErrorsFlusher to clear fatal errors in a messages pool before starting a new test. In case of a new test was started with a fatal errors we are restarting IDE.