[gui-test] check that path is selected
It could be dangerous to select path in async tree when the tree path may change (has been expanded for example) during mouse moving to the previously calculated location. So we are checking that finally selected row is intended and if is not calling clickPath again while it will be selected or attempts count will exhaust.