Allow Retry and Abort rebase from Untracked files error notification
To be able to alter the process that is triggering notification, it should be notified from the class responsible for the process.
Before this change, notification was notified from a standalone class thus not allowing to Retry or abort.
By this commit Notification is generated by the common GitUntrackedFilesHelper.java, but notified by the respective classes.
It allows to add actions to the notification respectively.

In future, this will allow to retry all processes triggering the Untracked files error, but for now implemented for rebase only.

This change is about to break API for 2 plugins:
com.google.cloud.tools.intellij.debugger.ProjectRepositoryValidator#unstash
com.urswolfer.intellij.plugin.gerrit.git.GerritGitUtil#cherryPick