PY-14358 Explicitly insert indent before the closing custom folding comment
that is the same as for the opening comment. Otherwise, it's impossible
to detect the necessary indentation in languages like Python, where the
block containing a comment is itself determined by this very indentation.

If it's wrong for some reason, the subsequent call of
CodeStyleManager.adjustLineIndent() should handle that in all other
languages.