project configuration: enable qualified module names by default
For a regular project it's enabled automatically if no module groups are configured. For projects imported from external system there is an option in 'Import Project' dialog which allows to either use explicit module groups (as before) or give qualified module names during importing and use them for grouping.
(IDEA-166061)