[vcs-log] index remembers commits that have full rename information saved
A special method is introduced to schedule reindexing for commits that do not have full rename information. It sends indexing requests with "reindex" flag. Requests with this flag do not modify myNumberOfTasks field: we do not consider them when checking if log is fully indexed or not.

(cherry picked from commit 61d2133)