Revert making Git VFS listeners asynchronous
This reverts commits 9ace3e7, a23fc196, 8b234f1, e286006

The fix was to move VFS events processing to a pooled thread.

But instances of GitRepositoryUpdater and GitUntrackedFilesHolder are
created per repository => in a multi-repo project there will be twice
as many pooled threads as there are repositories, which can lead
to performance degradation instead of improvement.

Reported in several issues e.g. in IDEA-179723