PY-24875: Filter messages in console, do not use InputFilter EP
With InputFilter API we can't check if are in test console and even
if it is Python console or not.
Affecting any console in Intellij-based product is not good idea
and may lead to performance problems.

See https://buildserver.labs.intellij.net/viewLog.html?buildId=22270794&tab=buildResultsDiv&buildTypeId=ijplatform_master_Idea_PerformanceTests