git: treat 'clone' as a read command IDEA-179899
In fact, clone is a write command, because it modifies index.lock;
however that index.lock is in new repository => it can't be accessed
by any other command at the same time.

If the clone is performed into a subfolder of the current project,
it still can't interfere with any other Git command, because the subfolder
must be empty (otherwise clone won't start).

This change allows however to call two clones into the same directory
simultaneously, and one of them will fail, but that doesn't look like
a serious issue and should anyway be solved by a more high-level IDEA-90211.