PY-24160 Do not give indent to a binary expression in parentheses
Use ContinuationIndent instead of ContinuationIndentWithoutFirst
for its operands and operators instead. This way we don't over-indent
a condition in parentheses when the opening bracket is on its own line.

Also test that binary expressions in other positions are not
over-indented thanks to the use of ContinuationWithoutFirst for them.

(cherry picked from commit fc30a55)