Deprecate PeriodicalTasksCloser and remove its usages
Requesting for a service or a component from a background thread
can lead to an "Already Disposed" exception.
PeriodicalTasksCloser was introduced to workaround this situation.

However, this approach is incorrect: instead of forking a thread,
catching exceptions and converting them to PCEs, the background
activity should be properly started under a dispose-aware ProgressIndicator
which throws PCEs automatically when a service is requested during
project disposal.