sm test runner: add optional ability to print only own content (i.e. skipping printing children) for root node if it's selected
Some test runners (e.g. jest, protractor) output test run details as human-readable plain text (sometimes it's pretty colored). In this case, a test failure is duplicated as it's printed as plain text (i.e uncaptured output for root node), and, also, it's printed as a state of failed test.