Gradle: request for extra models in the order of project resolver.
Some project model has to be requested after other models.
The original code use Set to save the model classes, but set doesn't
guarantee the iteration order. Change to LinkedHashSet to keep the
insertion order.
This ensures that the order of model request is consistent with the
order of ProjectResolver that model was defined with.