Fix VFS events processing in GitUntrackedFilesHolder
event.isValid() check, which was added in 816f311 to avoid EA-109103, was incorrect:
VFileCreateEvent treats event as not valid if the file already exists,
and it almost always already exists because events are process async.

Because of this, VFileCreateEvents were simply ignored: IDEA-180605.

isValid() check is reverted, and EA-109103 is fixed the other way:
by manually checking if the parent is valid.