[gui-test] build test classpath only for tests
After extending {IDE}-guitests modules classpath with {IDE}-main module we became experiencing a problem with an abnormally wide classpath. For example, GUI test for the WebStorm starts IDE with extra plugins from PhpStorm and JavaEE. That was happened because of very deep dependency tree for the test scope. Since this commit, we are building a classpath only for a production scope of testing module and adding a set of test classes (directory containing test sources or a jar). All test util methods should be placed in a production scope. The other problem is to resolve all macros (MAVEN_REPOSITORY and KOTLIN_BUNDLED) for generated classpath for a testing module by JPS. it is solved by finding a similar URLs in a built classpath for a testing module and replace the macros in JPSs classpath.