Honor studio.projectview when showing Project tool window.
After ag/2073215, ProjectView tool window default pane is Android isntead of
Project, but the change does not see what is the value of
studio.projectview.

This CL changes the default according to what studio.projectview is set. If
it is true it should be Project, if false (or not set), Android.

Bug: 67790043
Test: New Junit test to check default view honors system properties.

Change-Id: I379ce040393f43d2f014266ef157969e7dbbdc22