PY-25234 Don't give third-party priority to imports of modules from user skeletons
Use the same (britle) heuristic as for binary stubs here: if module is
in the root of python-skeletons directory consider it a part of stdlib.

It seems that we can spuriously return a stub from either
python-skeletons or binary_stubs for such stdlib modules, hence
the previous change wasn't enough to fix the problem completely.