Studio build: filter out empty folders
We use zip with default options to build Mac and Windows bundles, and
we end up storing empty, versioned folders for java-lib-model and
java-lib-model-builder. These two libraries are not shipped with the
offline repo. Typically patches rename these empty folders, but there
are some patches which simply delete them, such as:
https://android-devtools-staging.corp.google.com/no_crawl/studio/2.3.3.0_4069837/patches/AI-171.4141229-171.4159556-patch-mac.jar
... which means a further update will fail if it expects those empty
folders to be there, just to be deleted again.

As an aside, on Linux we build a .tar.gz which by default doesn't
include these empty directories.

Bug: 68162671
Bug: 63439489
Test: N/A
Change-Id: Ibb0014574826f541bdc9444f717d1f26445d7dd5