Wait for changelist manager update forever but be cancellation-friendly
Even 100 seconds can be not enough, if the CLM was about to refresh all
Git roots, together with the VFS refresh, indexing and compilation.

If the process doesn't finish ever, it indicates a bug in the CLM
refresh, which should be fixed separately.

But make the process cancellable, just in case.

Together with the previous commit rewriting the process of cherry-picking,
this should fix IDEA-143100.