PY-24265 Don't show preview for renamed functions in trivial cases
Because of a number of tricky scenarios related to dynamic features
of Python we used to always display a toolwindow with preview of found
usages for top-level definitions, however it doesn't make much sense
for functions and methods not called explicitly anywhere, such as
tests, or those functions that have just been written.

I added a new method forcesShowPreview(PsiElement, UsageInfo[]) in
RenamePsiElementProcessor that has the same purpose but is called
*after* the usage search has been performed, and thus it doesn't make us
search for usages over again in prepareRenaming().