[git] wait for process being cancelled and if wait is unsuccessful kill it
Git commands use soft kill for cancelling git processes: process is cancelled by sending a signal. Sending a signal does not terminate a process immediately, BaseOutputReader thread reads and processes the output until it is available. This is a problem for GitTextHandler.waitForProcess method, which needs to wait for process to finish, since all the clients expect it to do so. For example, VcsLogPersistentIndex does not want to start another indexing process until the first one is still alive.

So after calling destroyProcess, another waitFor with timeout is added. If timeout is expired, the process is destroyed in the old-fashioned way.