[gui-test] limit attempts for a mouse move to the desired point
SmartWaitRobot#moveMouse(c: Component, x: Int, y: Int) used an endless count of attempts to adjust a mouse cursor on the desired position. It has been created to avoid mistakes if a component has been changed its position during the mouse move. But it could disturb if using a Windows OS with a HighDPI display set and coordinates are not an integer. This commit limits a count of attempts of mouse position adjusting to 3 times by default.