jetCheck: smarter shrinking when item generation depends on previous items
For generators that generate sequences of commands with some internal model,
next commands might depend on the state of the model, modified by previous commands.

When previous commands are removed by shrinking, replaying generators would result
in a different models, and sampling from model elements would fail because
the model element counts have changed, and so should indices used for sampling.

Now we detect that we're given an int distribution that's different from the original,
and try to guess possible updated int values according to the new distribution.