PY-24412 Keep names of PyPI packages separately from the packaging settings
So as not to load the whole cache of available packages on EDT just
to get a timestamp for its invalidation in PyPackagesUpdater.

One simplification I made is that we no longer store the latest versions
of these packages in persisted settings. It would to much of a hassle
to keep it in sync then, and overall it's not that beneficial since
we're not able to get both package names and their versions in one
request due to the latest changes in PyPI API.