build scripts: allow removing version numbers from JAR names
When a library is converted to repository library, its version is included into its JAR name. So if we convert a library included into bootstrap classpath (names of such libraries are hardcoded in executables) we need to keep the old name to avoid changing executables each time the library is updated. Later we can automatically compose bootstrap classpath and this won't be needed.

(cherry picked from commit b832e6c)