Can't attach to local process on macOS with Python 3.6 (PY-26800)
In case of attaching to process, sdkHomePath appears from the running process. So there might be situation, when there is no Sdk, defined for this path, but executable exists, it's correct, and we should use for it for attaching. So, check for Sdk nullability is needed only in case when Console is enabled for run configuration.

(cherry picked from commit b389052)

(cherry picked from commit 673e927)