PY-21864 Support of ahead of time variable annotations
We search for them only in the same scope where the corresponding target
expression (of for loop, with statement or LHS of assignment with tuple
unpacking) is defined itself.

Another limitation is that this lookup is performed without proper
caching of this information on an affected target yet.

(cherry picked from commit d2eb0c6)

(cherry picked from commit 8d4e19f)