git tests: update CLM after setUp of GitHistoryUtilsTest
Although the test mainly tests how `git log` is requested and parsed,
GitHistoryUtils.getCurrentRevision() checks if there are local changes
to the file: if the file is locally renamed, `git log <old name>`
should be called, since the new name is not known to Git yet.

Therefore it is necessary to update the CLM before the test start,
otherwise it may capture the rename (which happens in the setUp)
as a local rename, which causes blinking.