git: log info instead of warn
No difference for production, but less spam in tests.

The error itself is valid, e.g. when trying to checkout a reference
which exists not in all repositories, collecting the pre-diff fails, but
it is easier and faster to try to call diff and fail instead of
calculating whether the reference exists (e.g. if it is a hash).