[gui-test] don't limit IDE start if debug mode is true
We limit the IDE start to 3 minutes by default. But it could be inconvenient if we are trying to debug test remotely and it is quite hard to catch the debugger listening on port in 3 minutes if other build process take for hour(s). So let's just don't limit process execution value if idea debug mode is true.