PY-24990 Add extra tests on more elaborate cases of annotations for cls/self
First, on matching type var with a union receiver type.

Though we're not able to match types like "Union[Type[A], Type[B]]" with
parametrized "Type[T]" to infer that T is Union[A, B] yet (should we?).

Second, on calling an annotated instance method on a class object
(when an argument for "self" is provided explicitly).