Each crash logged as a separate event.
Currently we only send aggregate information about
crashes that happened on user machines. While we have individual
stack trace hases, the timestamps of the events are not preserved -
crashes are sent with the timestamp of the SystemHealthMonitor ping
(which happens once an hour).

With this change, each crash will be logged with timestamp when it happened.
This will help with figuring out what happened just before the crash/exception.
As multiple crashes with the same stacktrace are aggregated, onyl the timestamp
of the first instance of such exception is reported.

Test: ExcetionRegistryTest.testStacktraceTimestamps
Change-Id: If67456f08b098425fbdeab94120bfeab593a298b