git: more correctly identify "local changes would be overwritten by"
If the "Local changes would be overwritten by merge/checkout" error
happens in one of repositories, we collect the changes which would
produce this error in subsequent repositories, to show the smart
checkout/merge dialog only once.

To identify these changes, the code used to get the diff between
the current branch and the branch to checkout:
    `git diff master..feature`
and then check for local changes among these files.
However, it is more correct to get the diff between the current working
tree and the branch to checkout:
    `git diff feature`
The difference is subtle, but new behavior is more correct.