git: spam less to the log
It is OK not to find some files by path, because the method is used
e.g. to get local changes from a `git diff <other branch>` output,
which contains file absent in the current working tree.

On the other hand, the log didn't prove itself as useful yet, so it's
OK to silence it. If we find that we should warn about the absent
file in certain cases, we can introduce a different method for that.