PY-12002 Restore providing keywords arguments in calls related to Django models
broken due to the recent update of KeywordArgumentCompletionUtil:
891c4bf405dad165a6d8c9daa0f15fcb72467107.

Namely, for classes with __init__ containing **kwarg, it's properly
passed to PyKeywordArgumentProviders. It broke down because reference in
a constructor call is resolved into both class and its constructor, but
only the type of the corresponding class is inferred. Also,
parameters of the actual function is used, not those supplied by
a type provider because the latter usually omit "self", and then
presense of **kwargs can be overlooked due to the way
KwArgParameterCollector handles them.

(cherry picked from commit ecdf075)

(cherry picked from commit 6ed2ff0)