Enable leak hunter for light projects
The issue reproduces only in bazel test runs with sharding enabled.
Due to timing, certain invocations scheduled with LaterInvocator (typicaly
with Application.invokeLater call) will expire before they are processed.
These invocations were still used as object roots for the LeakHunter.
Tests that consistently failed were
AndroidUtilsTest.testIsValidJavaPackageName and
AndroidUtilsTest.testValidateAndroidPackageName, but this is more of a
coincidence as there is nothing special in these tests.

VcsEventWatcher.projectOpened registered listener to rootsChanged message on
PORJECT_ROOTS topic. When roots has changed and the project has already been
disposed, a invokeLater runnable was scheduled. As the project got disposed
before this invocation was executed (and project.getDisposed() is the
condition for expiration), such invocation should not be used for object
roots.

Test: N/A
Change-Id: I1dd392c4d75f62de0fb165a3dfb3f25e20dc94c3