Allow overriding test.jvm.memory option from ant command line
After the 2017.3 merge, it is no more possible to override
the default value of "test.jvm.memory" when running ant tests
from the command line, for example:

ant -file build.xml "-Dtest.jvm.memory=-Xmx2400m -Dsun.io.useCanonCaches=false" test

This CL remove the hard-coded value from the gant script, and change the
default to 1024m in the groovy task. This ensures we have a default of
1024MB, while allowing the "-Dtest.jvm.memory=-Xmx....m" command line option
to override that default.

Bug: 70176320
Test: Manually ran "ant" command on Windows
Change-Id: Iba5f2e68df1e4b418fd04aa338b69d587d46a545