UI tests were stopping abruptly.
The build system registers uses a build logger
that throws an exception when an error message
is logged. This logger is used by JUnit. When
JUnit logs a failed test, the exception is thrown
from the build logger. This exception prevents
buildbot from performing the rest of the
test steps, such as processing the test result
files.

This change fixes this issue by telling JUnit
to not log errors. We already have loggers that
tell us when a particular test is breaking.

Bug: 69917400
Test: all existing tests
Change-Id: Ie25875723385c6db392f5e1521ce66413a500f4a