[vcs-log-graph] simplify print elements generation
We used to caclulate "simple row elements" (nodes and terminal arrows) and "down short edges" (edges from the current row to the next row) separately.
Then used those, plus "down short edges" for the previous row, to generate print elements.

The new approach is symmetrical: we walk through the elements in the row and for each edge get position in the both previous and next rows.
Using the positions we can generate up and down edges and arrows for this edge at the same time.

In order to later get rid of the intermediate classes, instead of returning a list, elements are fed to the special consumer.