[vcs-log] single task controller waits for the task on dispose
When SingleTaskController is disposed in background thread, it not only cancels the task but waits for it for a while.
In order to do so, a SingleTask abstraction is introduced, it has a progress indicator and a future inside.
SingleTask is cancelled prior to project dispose and SingleTaskController dispose. If SingleTaskController is disposed in background thread, it waits for the task to finish.