Optionally bundle UI tests in release build
This change introduces logic to allow the test scopes of modules to be
included in a PluginLayout, ensuring that test-scoped module and project
library dependencies are included.

Set the "bundle.ui.tests" system property to "true" (e.g. using the
--uitests flag to build_studio.sh) to bundle the uitest-framework plugin,
which contains both the framework and test code, as well as the test data.

This will enable UI tests to be run on the release version, with the aid
of framework additions from JetBrains' fork of our framework code.

Test: :)
Bug:
Change-Id: I2446cdf7f5e4642caa706ef574af87c5879dd248