[groovy] clean up GrOperatorExpression interface
Add `getOperation()`: `getOperation` differs from
`getOperationTokenType` in that the former should return actual
operation type which will be used for resolve/inference while the latter
is just a shorthand for obtaining token type of `getOperationToken`. In
case of binary expression it is the same as `getOperationTokenType`, in
case of assignment expression there corresponding operator type is
returned.

Pull down getOperationToken() and getOperationTokenType(): there two
methods aren't used via GrOperatorExpression and they should not.