runnerw: ignore insert sequences as none are provided
Allows to show error messages for some exit code with system messages containing %1.
For example, now message for 216 exit code:
===
runnerw.exe: CreateProcess failed with error 216: This version of %1 is not compatible with the version of Windows you're running. Check your computer's system information and then contact the software publisher.