[file-history] use proper root when indexing paths
When calculating affected paths in the commit, all parent directories up to repository root are considered affected. Previously any registered repository root was used, not necessarily the root containing specified commit. This led to a problem with submodule commits. A commit that changed submodule version was not considered to affect parent repository and did not show up in history for it.

IDEA-183938