Reveal directory structure for blaze plugins
This CL removes the module grouping associated with the Blaze plugins.
Instead, all the Blaze plugins will be shown as submodules of a new
"blaze-plugin" module. Since the blaze-plugin module has a content root
that points to tools/vendor/google3/blaze, IntelliJ displays the true
directory hierarchy of everything under that folder now.

Test: no production impact
Change-Id: I8fe635bbd7cf0d4bfff852a9130716b43238f6a6