git: fire an event if tag has changed even if something else also changed
GitRepository.update() fires a repositoryChanged() event only after
comparing the previous repository info with the one it has just read.
Plain tag operation can produce other kinds of events as well,
e.g. change the packed-refs file, but repo info comparison won't
show any difference, because we don't store and compare tags.

The fix will potentially fire two events (e.g. if a tag was deleted,
and a branch was changed), but that shouldn't be a problem.

This fixes IDEA-181273 and IDEA-184258.