[groovy] fix duplicate fields highlighting (IDEA-184971)
Before the change fields were fed to processor from
CollectClassMembersUtil, if there was a field with explicit visibility
modifiers, then only this field remained in the candidate list and fed
to the processor.

After the change the CompilationPhaseHint is used, and all code fields
are fed to the processor one by one. Also now processor returns after
the first duplicate is found, ignoring the rest.