PY-21034 PY-27907 PY-11031 Cython parsing made lang level independent
Cython pretends to work with both Python 2 and Python 3 constructs, so
Cython AST should not depend on project language level. To achieve this,
statement and function parsing were adopted to be language level
independent. However, there are still plenty of LanguageLevel
dependencies in Cython parsing code (see comments for CythonParser) that
are targets for refactoring.

To remove language level compatibility warnings Compatibility inspection
was turned off for Cython files.