Revert "Reveal directory structure for blaze plugins"
This reverts commit b30a3e7887e1b4b28f374d32e07521bb98c1ee96.

The earlier commit causes //tools/base/bazel:iml_to_build to generate an
iml_module reference inside the blaze plugin, and it generates it in an
incorrect format (doesn't understand nested workspaces). So this reverts
the CL until it can be re-introduced with the appropriate BUILD file
changes as well.

Test: N/A
Change-Id: Ibb8c3d87256cee00898c7ad30fadcd11b5f13aa7