[groovy] resolve fields before transformations
Before the change the `foo` reference inside of a closure was resolved
to a field. Now it's properly resolved to an accessor method (as done by
Groovy itself).
```
class A {
  def foo
  def bar = {
    foo
  }
}
```

Before the change the `foo` reference inside of a class was resolved to
a field even when it was qualified. Now it is resolved to an accessor
method.
```
class A {
  def foo
  def bar() {
    new A().foo
  }
}
```

This change also speeds up resolution as it stops when first field is
found, i.e. transformations and non-code processors will not be run.