[gui-test] expand jar's classpath to compose the module's classpath
Sometimes Windows doesn't accept long classpath as to command line limitations, than we need
to compose a fake module xml with a classpath list and pass it to the Kotlin compiler. Unfortunately,
if IDEA has been launched with a fake jar containing only a long classpath in its manifest, it doesn't
work. This commit expand all classpath[any-number].jar and pathing.jar to the set of unique jar and class
URLs and than we are composing a module xml to with a full classpath into a Kotlin compiler again.