Do not consider file as a target element for the very first leaf (GO-3661)
otherwise, since PsiFile is named element and its textOffset is 0,
the very first comment in a file will be considered as
named element during rename refactoring and won't be renamed at all.

Php rename tests were changed since they relied on the fact that
the a file is a target element for the very first "<?php" leaf and will be
renamed when caret is located at "<?php".

The initial issue was in Go support, JS test was added in order to have test for the change in platform.